app-sync-env.sh: Tool to loop over argocd app sync
Use this tool to force ArgoCD to complete an upgrade of the services that
it is monitoring. This is used when the new manifests from the gitops repo
contain new API versions in the CRDs its already monitoring and also
deployments that have webhooks that need the new API versions, and also some
new resources that use the new API versions.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>