A more find-grained approach to server-side-apply
The nnf-sos manifest is quite large because it has two CRDs that include
corev1/PodSpec or PodTemplateSpec, and because on some systems the
SystemConfiguration resource can be extensive. Rather than doing server-side
apply for the whole manifest, just do it for the resources that require it.

This seems to greatly simplify the verification steps that argocd has to do for
rolling-upgrade.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>