Reimplement unpack-manifest in python
Make it easier to add features to this tool.

It can now dynamically find new template resources and separate them and
create matching default resources.

The NOTEs that are displayed when the tool completes are now also recorded in
environment/$ENV/unpacking-notes.txt so that it is easy to refer back to them.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>