Fix nnf-dm coredump patches (#45)
Problem: The json patch for the nnf-dm controller deployment is using
indexes to patch the volumes, which is error prone. We recently had the
containers switch indices, so the patch was being applied to the wrong
container.

This changes the json patch to a strategic merge style patch for the
deployment, allowing the container name to be used to apply the patch.

Unfortunately, we cannot use a strategic merge for the patch to the
NnfDataMovementManager, since it is a CRD. Continue to use a JSON patch
in this case. It is less risky than the deployment since there are two
containers and we want to patch them both.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>