Make DWS upgrades go more smoothly (#49)
Note, for best behavior of DWS upgrades, these changes must be applied and
active in the gitops repo prior to doing the DWS upgrade.

Apply the SystemConfiguration with ServerSideApply=true and Replace=true. On
some systems this can be quite a large file, so it should be server-side. And
for DWS API upgrades we have to make it easy to get this past the DWS
conversion webhook so we deploy it with the equivalent of "kubectl replace".

Deploy the DWS webhook Deployment with Replace=true and Force=true. This allows
the previous replicaset to cleanup the previous DWS webhook pods prior to
starting the next replicaset. This addresses the issue where the previous pods
never go away on their own, which prevented the new pods from progressing
past Pending state.

Move the location of SystemConfiguration from the nnf-sos dir to the
site-config dir. In previous testing we've found that this behaves better when
deployed in this later stage. This is already active in the downstream gitops
repos, and was missed when syncing with the boilerplate.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>